🤖 Refactor(client): screenToGrid gridToScreen 함수 분리
